+++ include/core/ProgressAccuracy.h
@@ -7,12 +7,12 @@ class ProgressAccuracy : public ProgressMetric {
         static const string NAME;
         
         size_t correctPredictions;
-        vector<double> predictions;
+        vector<float> predictions;
 
     public:
         ProgressAccuracy(size_t);
         string getName() const override;
         void init() override;
-        void update(const Batch&, const Loss*, const Tensor&, double) override;
-        double calculate() const override;
+        void update(const Batch&, const Loss*, const Tensor&, float) override;
+        float calculate() const override;
 };
