+++ include/utils/Greyscale.h
@@ -3,18 +3,18 @@
 class Greyscale : public Scalar {
     private:
         // Constants
-        static const double MAX_GREYSCALE_VALUE;
+        static const float MAX_GREYSCALE_VALUE;
         void throwDataFormatError() const;
     
     public:
         // Methods
-        void fit(const vector<double>&) override;
+        void fit(const vector<float>&) override;
 
         Tensor transform(const Tensor&) const override;
         Tensor reverseTransform(const Tensor&) const override; 
 
-        vector<double> transform(const vector<double>&)const override;
-        vector<double> reverseTransform(const vector<double>&) const override;
+        vector<float> transform(const vector<float>&)const override;
+        vector<float> reverseTransform(const vector<float>&) const override;
 
         uint32_t getEncoding() const override;
 };
