+++ include/utils/TrainingUtils.h
@@ -9,13 +9,13 @@ using namespace std;
 class TrainingUtils {
     private:
         // Constants
-        static const double GRADIENT_THRESHOLD;
+        static const float GRADIENT_THRESHOLD;
         
     public:
         // Methods
-        static double getAccuracy(const vector<double>&, const vector<double>&);
-        static double clipDerivative(double);
-        static vector<double> getPredictions(const Tensor&);
-        static double getPrediction(const vector<double> &, size_t, size_t);
-        static double getRMSE(const Tensor&, const vector<double>&);
+        static float getAccuracy(const vector<float>&, const vector<float>&);
+        static float clipDerivative(float);
+        static vector<float> getPredictions(const Tensor&);
+        static float getPrediction(const vector<float> &, size_t, size_t);
+        static float getRMSE(const Tensor&, const vector<float>&);
 };
