+++ src/core/ProgressMAPE.cpp
@@ -21,20 +21,20 @@ void ProgressMAPE::update(
     const Batch &batch,
     const Loss *loss,
     const Tensor &outputActivations,
-    double batchTotalLoss
+    float batchTotalLoss
 ) {
     ProgressMetric::update(batch, loss, outputActivations, batchTotalLoss);
 
-    const vector<double> &outputFlat = outputActivations.getFlat();
-    const vector<double> &batchTargets = batch.getTargets();
+    const vector<float> &outputFlat = outputActivations.getFlat();
+    const vector<float> &batchTargets = batch.getTargets();
 
     size_t numBatchSamples = outputFlat.size();
-    double localMapeSum = 0.0;
+    float localMapeSum = 0.0;
     size_t localNonZero = 0;
 
     #pragma omp parallel for reduction(+:localMapeSum, localNonZero)
     for (size_t i = 0; i < numBatchSamples; i++) {
-        double actual = batchTargets[i];
+        float actual = batchTargets[i];
         if (actual != 0) {
             localNonZero ++;
             localMapeSum += abs(outputFlat[i] - actual)/actual;
@@ -45,7 +45,7 @@ void ProgressMAPE::update(
     numNonZeroTargets += localNonZero;
 }
 
-double ProgressMAPE::calculate() const {
+float ProgressMAPE::calculate() const {
     if (numNonZeroTargets == 0) {
         cerr << "Warning: All target values were 0. MAPE is undefined." << endl;
         return 0.0;
