+++ src/utils/FeatureEncoder.cpp
@@ -30,30 +30,30 @@ vector<bool> FeatureEncoder::getCategoricalCols(
     return isCategorical;
 }
 
-unordered_map<string, double> FeatureEncoder::encodeFeature(
+unordered_map<string, float> FeatureEncoder::encodeFeature(
     const vector<vector<string> > &featuresRaw,
     size_t colIdx
 ) {
     size_t numRows = featuresRaw.size();
-    unordered_map<string, double> encoding;
+    unordered_map<string, float> encoding;
     int nextIdx = 0;
 
     for (size_t i = 0; i < numRows; i++) {
         const string &val = featuresRaw[i][colIdx];
         if (encoding.find(val) == encoding.end()) {
-            encoding[val] = (double) nextIdx++;
+            encoding[val] = (float) nextIdx++;
         }
     }
 
     return encoding;
 }
 
-vector<unordered_map<string, double> > FeatureEncoder::encodeFeatures(
+vector<unordered_map<string, float> > FeatureEncoder::encodeFeatures(
     const vector<vector<string> > &featuresRaw,
     const vector<bool> &isCategorical
 ) {
     size_t numCols = featuresRaw[0].size();
-    vector<unordered_map<string, double> > encodings(numCols);
+    vector<unordered_map<string, float> > encodings(numCols);
 
     #pragma omp parallel for
     for (size_t j = 0; j < numCols; j++)  {
@@ -67,7 +67,7 @@ vector<unordered_map<string, double> > FeatureEncoder::encodeFeatures(
 
 vector<size_t> FeatureEncoder::getOffsets(
     const vector<bool> &isCategorical,
-    const vector<unordered_map<string, double> > &encodings,
+    const vector<unordered_map<string, float> > &encodings,
     size_t numCols,
     size_t &totalCols
 ) {
@@ -86,7 +86,7 @@ vector<size_t> FeatureEncoder::getOffsets(
 
 Tensor FeatureEncoder::getFeatures(
     const vector<bool> &isCategorical,
-    const vector<unordered_map<string, double> > &encodings,
+    const vector<unordered_map<string, float> > &encodings,
     const vector<vector<string> > &featuresRaw
 ) {
     size_t numRows = featuresRaw.size();
@@ -95,14 +95,14 @@ Tensor FeatureEncoder::getFeatures(
     size_t totalCols = 0;
     vector<size_t> offsets = getOffsets(isCategorical, encodings, numCols, totalCols);
     Tensor features({numRows, totalCols});
-    vector<double> &featuresFlat = features.getFlat();
+    vector<float> &featuresFlat = features.getFlat();
 
     #pragma omp parallel for
     for (size_t i = 0; i < numRows; i++)  {
         for (size_t j = 0; j < numCols; j++) {
             size_t offset = offsets[j];
             if (isCategorical[j]) {
-                unordered_map<string, double>::const_iterator it = encodings[j].find(featuresRaw[i][j]);
+                unordered_map<string, float>::const_iterator it = encodings[j].find(featuresRaw[i][j]);
                 if (it != encodings[j].end()) {
                     size_t catIdx = it->second;
                     featuresFlat[i * totalCols + offset + catIdx] = 1;
